server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Argana Phangquestian <arganaphangquestian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
